Dont restart checkout_flow for completed orders
No need to destroy all shipments or set order back to first checkout
state. Currently shipments for completed orders are managed by
OrderInventory#veriry

+++ core/app/models/spree/order.rb
@@ -491,7 +491,7 @@ module Spree
     # to delivery again so that proper updated shipments are created.
     # e.g. customer goes back from payment step and changes order items
     def ensure_updated_shipments
-      if shipments.any?
+      if shipments.any? && !self.completed?
         self.shipments.destroy_all
         self.update_column(:shipment_total, 0)
         restart_checkout_flow
